✨ anomalist: stop using mock data (#3410)
* ✨ anomalist: stop using mock

* style

* ✨ anomalist: stop using mock data

* re-order mock data

* replace mock data with real data

* discard df if all-zero
